Add license mechanism (auto selected license and license file)
By law it isn't really possible to put a specific license on public
sector information as the laws define the general terms. This means
we will save people some time and automatically choose the general
terms description and at that to the license file which we'll
include. This means license_group_url should be set to:
https://domain.is/licenses/iceland.json